Added program runtime control variables.
global_time will be the incremented variable that shows what second of runtime
the simulator is on. This is needed to move packets from the wire to a router,
from the incoming queue on the router to the outgoing queue on the router, and
from the router back onto the wire.
SECONDS_TO_RUN will control the total number of seconds that the program will
run for the duration of the simulation.