Controls the Incoming Connections on the Router
This connection class is for incoming packets. The first step is to receive a
packet then check if the destination of the packet is the receiving router. If
the router is the receiver the packet is freed from memory and the counter for
the number of packets received is incremented. Otherwise the router will attempt
to enqueue the packet on the incoming queue. If there is room on the queue the
packet will be added to the incoming queue otherwise it will be dropped from the
router and the dropped packets counter will be incremented. When it is added to
the queue the time that it takes to process the packet will be added to the
packet's variable that says when the next process occurs and the packet will
wait to be switched to the outgoing packet queue. After the global_time gets to
a point where the packet can be switched, the packet will be returned to the
router to be switched to the outgoing queue.