Controls the Outgoing Connections on the Router
This connection class is for outgoing packets. The first step is to receive a
packet from the incoming packet queue then put it on the outgoing connection
queue. If there is no room on the queue then the packet is dropped. Otherwise
the packet is enqueued, the time to switch is calculated, then the packet awaits
transmission to the wire. After the global_time is to a point where the packet
can be switched, the packet will be sent to the router where it will be wrapped
and send out onto the wire.